No chieldList
I was with a problem where I didn't was able to select anything and maintain selected when the options of autochanges was activated (maybe too many extensions), I find out that for me was the argument chieldList. I removed and the this error stoped. If I doesn't update this, it's because didn't cause any other problem.