Reverted changes
Wasn't activating the auto-detect page changes because of it. But I find out the problem that impossibilite me to select things, it's because of the class with the id: tooltip-container; delecting that, fixed my problem. I'm using a script in TamperMonkey with jQuery to delete this id from the website, I don't know how to implement this here. Maybe, I'll try it just for fun.